Change error bars to points with ranges
The problem with error bars is that in places where there is no range of
estimate, it looks like we have better data than we do. Actually, in
those places the data is probably worse, because we don't know what the
range is. So using geom_pointrange lets us see the range as additional
information. As a plus, we also get dots at the data points, which makes
the underlying data more obvious.